Fix: Correct import paths and add missing test file after ai-engine refactor
