+++ ai-engine/__main__.py
@@ -2,7 +2,7 @@
 Entry point for python -m modporter.cli
 """
 
-from src.cli.main import main
+from cli.main import main
 
 if __name__ == '__main__':
     main()
