+++ ai-engine/tests/integration/test_basic_integration.py
@@ -1,0 +1,4 @@
+import pytest
+
+def test_basic_integration_example():
+    assert True
